Node-red flow expanded to support inbound messages
The Twine functions that are made available in the provided Javascript are received in this flow, which has examples of how to capture and respond to the data, as well as a minimal example story in the HTTP node.

This flow can be deployed and makes a game available at http://localhost:1880/game (use a different hostname if accessing it from another machine on  the same network, natch).