Example Twine game
This game needs to be placed in the Twine library (menu Twine→Show Library) and opened from within Twine.

In the game editor, menu→Publish To File will export a HTML file. Copy the contents of this file into your clipboard, and paste it into the `game code` node in the provided Node-red flow.

On a mac, a quick way to get the file contents into the clipboard is to open a terminal, type 'cat ', followed by a space, then dragging-and-dropping the HTML file into the terminal (to get its path), then typing a space, then typing `| pbcopy'.

You should have a line that looks like

`cat /Users/yourname/Desktop/mygame.html | pbcopy`

Press return. 💥! Your game is in your clipboard, ready to be pasted.